requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.9
# Django==4.2.16
django-cors-headers==4.5.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
filelock==3.16.1
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.5
openapi-codec==1.3.2
packaging==24.1
platformdirs==4.3.6
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.27.0
